requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework
cx_Oracle>=8.2
djangorestframework-simplejwt
django-filter
django-cors-headers
pre-commit
pandas
openpyxl==3.0.9
django-pandas==0.6.6
xlsxwriter==3.0.2
drf-extensions
py-mini-racer==0.6.0
jsbeautifier
